Revert "[BENCH-471] Move test values to appropriate fixtures classes (#1353)"
This reverts commit abfbce7a4feb82fcb246c23b07b8e1848b151192.
