+++ service/src/test/java/bio/terra/workspace/common/fixtures/WorkspaceFixtures.java
@@ -1,5 +1,7 @@
 package bio.terra.workspace.common.fixtures;
 
+import static bio.terra.workspace.common.utils.MockMvcUtils.DEFAULT_USER_EMAIL;
+
 import bio.terra.common.exception.ApiException;
 import bio.terra.common.exception.NotFoundException;
 import bio.terra.common.exception.UnauthorizedException;
@@ -30,7 +32,6 @@ import org.testcontainers.shaded.org.apache.commons.lang3.RandomStringUtils;
 
 public class WorkspaceFixtures {
 
-  public static final UUID WORKSPACE_ID = UUID.fromString("00000000-fcf0-4981-bb96-6b8dd634e7c0");
   public static final String WORKSPACE_NAME = "TestWorkspace";
   public static final ApiProperty TYPE_PROPERTY =
       new ApiProperty().key(Properties.TYPE).value("type");
@@ -43,10 +44,8 @@ public class WorkspaceFixtures {
   public static final String DEFAULT_SPEND_PROFILE = "wm-default-spend-profile";
   public static final SpendProfileId DEFAULT_SPEND_PROFILE_ID =
       new SpendProfileId(DEFAULT_SPEND_PROFILE);
-  public static final String DEFAULT_USER_EMAIL = "fake@gmail.com";
-  public static final String DEFAULT_USER_SUBJECT_ID = "subjectId123456";
   public static final SamUser SAM_USER =
-      new SamUser(DEFAULT_USER_EMAIL, DEFAULT_USER_SUBJECT_ID, new BearerToken("token"));
+      new SamUser("example@example.com", "123ABC", new BearerToken("token"));
   public static final ApiException API_EXCEPTION = new ApiException("error");
   public static final NotFoundException NOT_FOUND_EXCEPTION = new NotFoundException("not found");
   public static final UnauthorizedException UNAUTHORIZED_EXCEPTION =
