+++ service/src/test/java/bio/terra/workspace/service/resource/ResourceLineageUtilsTest.java
@@ -1,6 +1,6 @@
 package bio.terra.workspace.service.resource;
 
-import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.DEFAULT_USER_EMAIL;
+import static bio.terra.workspace.common.utils.MockMvcUtils.DEFAULT_USER_EMAIL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import bio.terra.workspace.common.BaseUnitTest;
